+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - **Added new `HeaderBreadcrumbs` and `HeaderButton` component** for use with the Header `breadcrumbs` prop.
 - **Added date filter for the `Table`.**
 
+### Fixes
+
+- **Fixed `Table` bug where filter actions would also trigger sorting.**
+
 ## [1.4.1]
 
 `2021-05-17`